requirements.txt
requests
plyer
google-generativeai
beautifulsoup4
win10toast_click
cx_Freeze